requirements.txt
elasticsearch[async]
motor
black
pyyaml
